be/pyproject.toml
[project]
name = "spectra"
version = "0.1.0"
description = "A library for logging and analyzing LangGraph workflows."
authors = [
    {name = "Ryan",email = "rqn.pham@gmail.com"},
    {name = "Mizan",email = "mizantompkins@gmail.com"},
    {name = "Hunter",email = "hunter.nguyen@sjsu.edu"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph>=0.4.5,<0.5.0",
    "langchain-openai>=0.3.17,<0.4.0",
    "langchain>=0.3.25,<0.4.0",
    "python-dotenv>=1.1.0,<2.0.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["spectra*"]
